requirements.txt
# Core Dependencies
streamlit>=1.31.0
feedparser>=6.0.10
pandas>=2.0.0
requests>=2.31.0

# XML Parsing
lxml>=5.0.0

# Environment Variables
python-dotenv>=1.0.0

# Progress Bars
tqdm>=4.66.0

# Retry Logic
tenacity>=8.2.0

# LLM API Clients
openai>=1.12.0
anthropic>=0.18.0
